# docs/hld/openapi.yaml
...
openapi: 3.1.0
info:
  title: System-112 / DDS training simulator — backend API
  version: 1.0.0
  summary: Commands and resources for the local emergency-response training simulator.
  description: |
    REST command/resource API under `/api/v1` (D8). Realtime session updates travel over the
    WebSocket channel specified in `docs/hld/40-realtime-protocol.md`; raw microphone audio
    travels over LiveKit and never over this API (SPEC §34).

    Authorisation of every command is two-gated (D8): the caller's participant record must be
    assigned to the session's active `RoleStage`, **and** the stage's `RoleModule` must permit
    the action in the current stage state. A failure of the first gate is `403 FORBIDDEN_FOR_ROLE`;
    of the second, `409 ACTION_NOT_AVAILABLE`. An illegal state-machine trigger is
    `409 INVALID_TRANSITION`.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local docker-compose backend (SPEC §36).
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sessions:
    post:
      tags: [sessions]
      operationId: createSession
      summary: Create a simulation session (INSTRUCTOR / ADMIN).
      description: |
        Creates one `SimulationSession`, one `Incident` and one `RoleStage` per `role_chain`
        entry, instantiates `WorldTruth` and `CallerBelief` from the version, then fires the
        session machine trigger `validate` (`CREATED → READY`). A `role_chain` of `[DDS]` under
        `SINGLE_ROLE` or `ASSESSMENT` without `expected_response.prefab_handoff` is refused with
        `409 PREFAB_HANDOFF_REQUIRED` (D6, §10.10).
      x-action: '-'
      x-emits: [SESSION_CREATED, ROLE_STAGE_STARTED]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCreateRequest' }
      responses:
        '201':
          description: The created session in state `READY` (or `CREATED` if validation deferred).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      tags: [sessions]
      operationId: listSessions
      summary: List sessions — mine, or all for an instructor.
      description: |
        `scope=MINE` returns the sessions the caller participates in or created.
        `scope=ALL` is permitted for `INSTRUCTOR` and `ADMIN` only; a `TRAINEE` asking for
        `ALL` gets `403 FORBIDDEN_FOR_ROLE`.
      x-action: '-'
      x-emits: []
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [MINE, ALL]
            default: MINE
        - name: state
          in: query
          required: false
          description: Filter by session state.
          schema: { $ref: '#/components/schemas/SessionState' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Sessions, newest first.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SessionListItem' }
                  total: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/sessions/{session_id}/abort:
    post:
      tags: [sessions]
      operationId: abortSession
      summary: Abort the session (INSTRUCTOR / ADMIN).
      description: |
        Fires `abort` on the session machine and `abort_stage` on the active `RoleStage`.
        Simulation data is preserved: the event log is closed, never deleted (SPEC §39,
        §42 test 14).
      x-action: abort
      x-emits: [SESSION_ABORTED, STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AbortSessionRequest' }
      responses:
        '200':
          description: The session in state `ABORTED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/sessions/{session_id}/operator/stage/complete:
    post:
      tags: [operator]
      operationId: completeOperatorStage
      summary: Complete the Operator 112 stage.
      description: |
        Fires `complete_stage` (`HANDED_OFF → STAGE_COMPLETED`, guard `guard_call_ended`).
        The session machine then fires `begin_role_transition` when the `role_chain` has a next
        entry, or `complete` when it does not.
      x-action: complete_stage
      x-emits: [ROLE_STAGE_COMPLETED, STAGE_STATE_CHANGED, ROLE_TRANSITION_STARTED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The session after stage completion.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ------------------------------------------------------ role transition ----
...
  /api/v1/sessions/{session_id}/dds/work-item:
    get:
      tags: [dds]
      operationId: getDdsWorkItem
      summary: The incoming work item — built only from the HandoffSnapshot.
      description: |
        Permission `VIEW_HANDOFF`. The response schema `DdsWorkItem` contains **no field
        sourced from WorldTruth** and no field sourced from the live `OperatorCard`: it is
        assembled from `handoff_snapshots` and `dds_assignments` alone. The application service
        behind it is constructed without a world-truth repository, so the data is unreachable
        rather than merely unrequested (D3, SPEC §10, §42 test 3). If the 112 operator omitted
        a critical fact, the field is absent here and stays absent.
      x-action: '-'
      x-emits: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The work item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DdsWorkItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sessions/{session_id}/dds/acknowledge:
    post:
      tags: [dds]
      operationId: acknowledgeDdsAssignment
      summary: Acknowledge the work item.
      description: |
        Permission `ACKNOWLEDGE_ASSIGNMENT`; fires `acknowledge` (`RECEIVED → ACKNOWLEDGED`,
        guard `guard_participant_assigned_to_stage`).
      x-action: acknowledge
      x-emits: [DDS_ACKNOWLEDGED, STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The new DDS stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DdsStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/resources:
    get:
      tags: [dds]
      operationId: listDdsResources
      summary: The resource board for this assignment.
      description: |
        Permission `VIEW_RESOURCE_BOARD`. Scenario-defined resources only; the ETA numbers are
        `EtaProfile` values read verbatim by `ScenarioDefinedEta` (D7) — no external maps API
        is involved (SPEC §11).
      x-action: '-'
      x-emits: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: service_type
          in: query
          required: false
          schema: { $ref: '#/components/schemas/ServiceType' }
        - name: status
          in: query
          required: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/ResourceStatus' }
      responses:
        '200':
          description: The resource board.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/EmergencyResourceView' }
                  total: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sessions/{session_id}/dds/resources/selection/open:
    post:
      tags: [dds]
      operationId: openDdsResourceSelection
      summary: Open the resource-selection screen (additive, E9).
      description: |
        (additive, E9) Fires `open_resource_selection`
        (`ACKNOWLEDGED → RESOURCE_SELECTION`, unguarded; and
        `DISPATCHED → RESOURCE_SELECTION`, guard `guard_additional_dispatch_allowed`).

        §10.9 lists `open_resource_selection` as an available action of both states but §11 gave
        it no endpoint, so the DDS console had a button that could not be pressed. This is that
        endpoint, and it mirrors the operator's `beginHandoffPreparation` / `backToInterview`
        pair exactly: no request body, and the new `DdsStageView` in reply.
      x-action: open_resource_selection
      x-emits: [STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The new DDS stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DdsStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/resources/selection/cancel:
    post:
      tags: [dds]
      operationId: backToDdsAcknowledged
      summary: Leave the resource-selection screen without selecting (additive, E9).
      description: |
        (additive, E9) Fires `back_to_acknowledged` (`RESOURCE_SELECTION → ACKNOWLEDGED`, guard
        `guard_no_resource_selected` — stepping back may lose no work, so it is refused with
        `409 INVALID_TRANSITION` while a resource is `SELECTED`).

        The mirror of `openDdsResourceSelection`, and the DDS counterpart of the operator's
        `backToInterview`.
      x-action: back_to_acknowledged
      x-emits: [STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: The new DDS stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DdsStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/resources/select:
    post:
      tags: [dds]
      operationId: selectDdsResource
      summary: Select one resource.
      description: |
        Permission `SELECT_RESOURCES`. Fires the resource transition `select`
        (`AVAILABLE → SELECTED`, guard: the assignment is in `RESOURCE_SELECTION` and the
        resource is inside its availability window — otherwise `409 RESOURCE_UNAVAILABLE`).
      x-action: select_resource
      x-emits: [RESOURCE_SELECTED, RESOURCE_STATUS_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceSelectionRequest' }
      responses:
        '200':
          description: The new DDS stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DdsStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/resources/deselect:
    post:
      tags: [dds]
      operationId: deselectDdsResource
      summary: Deselect one resource.
      description: |
        Fires `deselect` (`SELECTED → AVAILABLE`, guard: the resource is not yet dispatched).
      x-action: deselect_resource
      x-emits: [RESOURCE_DESELECTED, RESOURCE_STATUS_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceSelectionRequest' }
      responses:
        '200':
          description: The new DDS stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DdsStageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/resources/dispatch:
    post:
      tags: [dds]
      operationId: dispatchDdsResources
      summary: Dispatch every currently selected resource.
      description: |
        Permission `DISPATCH_RESOURCES`. Fires the assignment trigger `dispatch`
        (`RESOURCE_SELECTION → DISPATCHED`) the first time, and `dispatch_additional` (a
        self-transition from `EN_ROUTE` / `ARRIVED` / `WORKING`) afterwards — extra units may be
        added without losing progress. Guard `guard_at_least_one_selected_available`.

        `RESOURCE_DISPATCHED` carries `capabilities_union`, `eta_seconds_by_resource` and
        `service_type_by_resource` (additive, E9) so the `RESOURCE_SELECTION` evaluator needs no
        resource-table lookup (D5). `service_type_by_resource` maps each dispatched
        `resource_id` to **that unit's own** `service_type`: a unit whose service received no
        handoff leg is attached to the primary leg, so an evaluator may never infer a unit's
        service from the assignment it hangs on.
      x-action: dispatch
      x-emits: [RESOURCE_DISPATCHED, RESOURCE_STATUS_CHANGED, STAGE_STATE_CHANGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispatchRequest' }
      responses:
        '200':
          description: The dispatch result and the new DDS stage view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DispatchResultView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/status-updates:
    post:
      tags: [dds]
      operationId: sendDdsStatusUpdate
      summary: Send an incident status update.
      description: |
        Permission `SEND_STATUS_UPDATE`. Not a stage transition: it is available in every state
        from `ACKNOWLEDGED` to `RESOLVED` and is what `REQUIRED_STATUS_UPDATE` scoring reads.
      x-action: send_status_update
      x-emits: [DDS_STATUS_UPDATE_SENT]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusUpdateRequest' }
      responses:
        '201':
          description: The recorded status update.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusUpdateView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /api/v1/sessions/{session_id}/dds/notifications:
    get:
      tags: [dds]
      operationId: listNotifications
      summary: Notifications addressed to the caller's role.
      description: |
        Read from the `notifications` table, materialized from `NOTIFICATION_CREATED` /
        `NOTIFICATION_ACKNOWLEDGED`. Filtered to `audience_role == the caller's role`; the
        operator console calls the same endpoint.
      x-action: '-'
      x-emits: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: unacknowledged_only
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Notifications, newest first.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/NotificationView' }
                  total: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sessions/{session_id}/dds/notifications/{notification_id}/acknowledge:
    post:
      tags: [dds]
      operationId: acknowledgeNotification
      summary: Acknowledge one notification.
      description: Permission `ACKNOWLEDGE_NOTIFICATION`. Held by both `Operator112Module` and `DDSModule`.
      x-action: acknowledge_notification
      x-emits: [NOTIFICATION_ACKNOWLEDGED]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: notification_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The acknowledged notification.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sessions/{session_id}/dds/radio-messages:
    get:
      tags: [dds]
      operationId: listRadioMessages
      summary: Radio traffic addressed to the caller's role.
      description: |
        Radio messages have **no table**: they are read from `session_events`
        (`RADIO_MESSAGE_CREATED`) through a read model built from the log
        (`docs/hld/20-db-schema.md` §20.1). Filtered to `to_role == the caller's role`.
      x-action: '-'
      x-emits: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: after_seq_no
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: Radio messages in log order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, last_seq_no]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RadioMessageView' }
                  last_seq_no: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT minted by `loginUser`; the secret comes from the environment (D8, SPEC §41).
...
  schemas:
...
    SessionCreateRequest:
      type: object
      additionalProperties: false
      required: [scenario_version_id, session_mode, participants]
      properties:
        scenario_version_id: { type: string, format: uuid }
        session_mode: { $ref: '#/components/schemas/SessionMode' }
        participants:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ParticipantAssignment' }
        session_seed:
          type: [string, 'null']
          description: |
            Overrides `ScenarioVersion.deterministic_seed`. Recorded in the `SESSION_CREATED`
            payload, so an identical seed with identical timestamped actions reproduces an
            identical world-event stream (D7, SPEC §42 test 7).
        time_scale:
          type: number
          minimum: 0.1
          maximum: 10
          default: 1
          description: |
            Multiplier applied to simulated time only. It scales world-event scheduling and ETA
            countdowns; it never scales the audio path, which is realtime by nature.
...